+++ src/test/test_vector.c
@@ -6,14 +6,14 @@
 /*   By: achappui <achappui@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/02 15:20:24 by achappui          #+#    #+#             */
-/*   Updated: 2024/08/02 16:17:50 by achappui         ###   ########.fr       */
+/*   Updated: 2024/08/03 07:09:25 by achappui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libla.h"
 #include "libft.h"
 
-t_vec4	test_vec4(double x, double y, double z, double w)
+t_vec4	test_vec4(float x, float y, float z, float w)
 {
 	t_vec4	vector;
 
@@ -61,7 +61,7 @@ void	vec4_cross_product_test()
 void	vec4_div_test()
 {
 	t_vec4	v_test1;
-	double	scalar;
+	float	scalar;
 	t_vec4	v_expected;
 	t_vec4	res;
 
@@ -78,7 +78,7 @@ void	vec4_div_test()
 void	vec4_mul_test()
 {
 	t_vec4	v_test1;
-	double	scalar;
+	float	scalar;
 	t_vec4	v_expected;
 	t_vec4	res;
 
@@ -113,8 +113,8 @@ void	vec4_dot_product_test()
 {
 	t_vec4	v_test1;
 	t_vec4	v_test2;
-	double	v_expected;
-	double	res;
+	float	v_expected;
+	float	res;
 
 	v_test1 = test_vec4(1, 2, 3, 6);
 	v_test2 = test_vec4(-2, 1, -2, 3);
@@ -161,8 +161,8 @@ void	vec4_inv_test()
 void	vec4_magnitude_test()
 {
 	t_vec4	v_test;
-	double	v_expected;
-	double	res;
+	float	v_expected;
+	float	res;
 
 	v_test = test_vec4(1, -2, 3, 3.31662479036);
 	v_expected = 5;
@@ -188,7 +188,7 @@ void	vec4_normalize_test()
 		ft_printf("vec4_normalize_test: KO\n");
 }
 
-t_vec3	test_vec3(double x, double y, double z)
+t_vec3	test_vec3(float x, float y, float z)
 {
 	t_vec3	vector;
 
@@ -252,7 +252,7 @@ void	vec3_add_test()
 void	vec3_mul_test()
 {
 	t_vec3	v_test1;
-	double	scalar;
+	float	scalar;
 	t_vec3	v_expected;
 	t_vec3	res;
